requirements.txt
fastapi
uvicorn
python-multipart
pillow
google-genai

python-dotenv
